Refactor planning and tool management features
- Moved Planner, a class for managing goal-oriented planning and execution of tool steps into its own file.

- Moved ToolRegistry, a class to manage tool registration and invocation into its own file.

- Brokeout / separated various file system tools (file_list, file_metadata, summarize_file, grep_files, find_file) for file operations and analysis into their own file.
